% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeColors.R
\name{attributeColors}
\alias{attributeColors}
\title{Attribute colors to network clusters}
\usage{
attributeColors(network, nb.max.colors = 12, palette = "Paired",
  lvl = "lvl1", colname = "color", sh.grey = FALSE,
  other.color = grey(0.5))
}
\arguments{
\item{network}{the network data.frame, usually issued from
\code{\link{readInfomapTree}} or getSitesTable or getSpeciesTable.}

\item{nb.max.colors}{a numeric <= 12 specifying the maximum number of
clusters to be colored.}

\item{palette}{a character string specifying the palette to use. See
\code{\link[RColorBrewer]{brewer.pal}}}

\item{lvl}{a character string specifying the level to which you want to
attribute colors}

\item{colname}{a character string specifying how to name the color column}

\item{sh.grey}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, then all clusters
beyond \code{nb.max.colors} will be attributed distinct shades of grey. If
\code{FALSE}, they will all have the color specified at \code{other.color}}

\item{other.color}{a character string or function specifying the color to
attribute to clusters beyond \code{nb.max.colors}}
}
\value{
The same data.frame as \code{network} with an additional color
column
}
\description{
This function adds a color column to a network data.frame using RColorBrewer
palettes. A maximum number of clusters to colorize can be specified to only
highlight the most important clusters.
}
\examples{



}
\author{
Boris Leroy \email{leroy.boris@gmail.com}
}
