% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkmetrics.R
\name{clusterMetrics}
\alias{clusterMetrics}
\title{Calculate species and site metrics according to clusters}
\usage{
clusterMetrics(
  db,
  network = NULL,
  site.field = colnames(db)[1],
  species.field = colnames(db)[2],
  site.area = NULL,
  level = "lvl1",
  stats.per.cluster = FALSE,
  ...
)
}
\arguments{
\item{db}{a data.frame with at least two columns, site and species}

\item{network}{a network data.frame containing cluster information,
typically obtained from \code{\link{readInfomapTree}}}

\item{site.field}{name or number of site column}

\item{species.field}{name or number of species column}

\item{site.area}{\code{NULL} or a data.frame with two columns: \code{name}
(site name, for correspondence with \code{network}) and \code{area}
(contains the area of each site). If \code{NULL}, area-based
metrics will not be calculated (see details).}

\item{level}{a character string specifying the column/level containing
clusters}

\item{stats.per.cluster}{experimental, a boolean indicating if site stats
should be calculated from a focal cluster perspective}
}
\value{
A list with 3 elements
\itemize{
\item{\code{region.stats}: a data.frame containing basic region stats (number
 of sites, area)}
\item{\code{species.stats}: a data.frame containing species-level metrics}
\item{\code{site.stats}: a data.frame containing site-level metrics}
}
}
\description{
This function calculate a series of metrics for species and sites with
respect to their characteristic clusters.
}
\details{
This function calculates several metrics for species and sites. First,
species-level metrics are calculated: affinity, fidelity, Indicator Value
and Dilution Value. Second, site-level metrics are calculated on the basis of
the metrics of species occurring within sites.

\bold{Species-level metrics}

Species-level metrics are calculated on the basis of the distribution range
of species compared to the size of clusters in which they occur. Hence, they
are supposed to be calculated on areas. However, sometimes areas are not
available, so we implemented an occurrence-based version of these metrics,
i.e. based on counts of species occurrences rather than based on distribution
range size.

All metrics have an occurrence-based (always prefixed with \code{Occ.}) and
area-based version. Occurrence-based metrics are always calculated; area-
based metrics are only calculated if argument \code{site.area} is filled
(see examples).


\itemize{
\item{Affinity of species to their region, \emph{Ai}

\ifelse{html}{\out{A<sub>i</sub> = R<sub>i</sub> / Z}}{\deqn{A_i=R_i/Z}}
where \emph{Ri} is the occurrence/range size of species \emph{i} in its associated
cluster, and \emph{Z} the total size (number of sites or total area) of the
the cluster

A high affinity means that the species is occupying most sites/areas of its
associated cluster.
}

\item{Fidelity of species to their region, \emph{Fi}

\ifelse{html}{\out{F<sub>i</sub> = R<sub>i</sub> / D<sub>i</sub>}}{\deqn{F_i=R_i/D_i}}
where \emph{Ri} is the occurrence/range size of species \emph{i} in its associated
cluster, and \emph{Di} is its total occurrence/range size.

A high fidelity means that the species is not present in other clusters than
their associated one.
}

\item{Indicator Value of species, \emph{IndVal}

\ifelse{html}{\out{IndVal = F<sub>i</sub> * A<sub>i</sub>}}{\deqn{IndVal=F_i A_i}}
where \emph{Fi} is the Fidelity of species \emph{i} to its associated
cluster, and Ai is its Affinity to its cluster.

The IndVal will have a high value for species that occupy most of their
associated cluster (high affinity) and are not present in any other
cluster (high fidelity).}


\item{Dilution Value of species for their region, \emph{IndVal}

\ifelse{html}{\out{DilVal = (1 - F<sub>i</sub>) * A<sub>i</sub>}}{\deqn{DilVal=(1-F_i) A_i}}
where \emph{Fi} is the Fidelity of species \emph{i} to its associated
cluster, and Ai is its Affinity to its cluster.

The DilVal substitues Fidelity (Fi) for Infidelity (1-Fi). A species
overlapping with much of its associated cluster (high affinity) but mostly
distributed in other clusters (high infidelity) will have a high DilVal.
}
}

\bold{Site-level metrics}

Site-level metrics are calculated on the basis of indices of species occurring
in sites.

\itemize{
\item{Robustness of cells, \emph{Rg}

\ifelse{html}{\out{R<sub>g</sub> = sum(IndVal<sub>i=1..C</sub>) - sum(DilVal<sub>j=1..N</sub>)}}{\deqn{R_g = \sum_{i=1}^C IndVal_i - \sum_{j=1}^N DilVal_j}}

where \emph{C} is the number of characteristic species of the cluster in
site \emph{g}, and \emph{N} is the number of non-characteristic species.

Grid cells containing a large number of
highly indicative species and a few species with low dilution
values will have high biogeographical robustness and will be
less vulnerable to changing their zooregion assignation due to
e.g. changes in the dataset (addition/removal of species as new samplings
are included in the dataset or corrections are made), or species introductions,
and extinctions.
}
\item{Relative robustness of cells, \emph{RRg}.
In case of large differences in size or sampling intensities between sites,
the robustness of cells can be overly influenced by species richness. Hence,
we implemented a "Relative" version of the robustness index:

\ifelse{html}{\out{RR<sub>g</sub> = R<sub>g</sub> / S}}{\deqn{RR_g = R_g / S}}

where \emph{S} is the total species richness of site \emph{g}.

}}
}
\examples{



}
\author{
Boris Leroy \email{leroy.boris@gmail.com}
}
