% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustersBelowLimit.R
\name{clustersBelowLimit}
\alias{clustersBelowLimit}
\title{Identify clusters that do not have a minimum number of nodes}
\usage{
clustersBelowLimit(df, column, limit, remove.clusters = FALSE)
}
\arguments{
\item{df}{the network data.frame, usually issued from
\code{\link{readInfomapTree}}}

\item{column}{column name of the level that you want to analyse}

\item{limit}{a numeric indicating the minimum number of nodes that separates
clusters above the limit from below the limit. Clusters must have a number of
nodes strictly superior to \code{limit} to be above.}

\item{remove.clusters}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, then
the function will return a data.frame without clusters below \code{limit}
\code{\link{readInfomapTree}}}
}
\value{
If \code{remove.clusters = TRUE}, a subset of \code{df} will be
returned without all nodes below the specified \code{limit}
}
\description{
This function identifies clusters that do not have a specified minimum number
of nodes, and can optionally remove them from the network data.frame.
}
\examples{



}
\author{
Boris Leroy \email{leroy.boris@gmail.com}
}
