% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTables.R
\name{getSiteTable}
\alias{getSiteTable}
\title{Get the data.frame of sites from a bipartite data.frame}
\usage{
getSiteTable(db, site.field = 1, network, drop.levels = TRUE)
}
\arguments{
\item{db}{the database that was used to construct the network.}

\item{site.field}{a character string giving the name of site column in
\code{db}}

\item{network}{the network data.frame, usually read from
\code{\link{readInfomapTree}}}

\item{drop.levels}{if \code{TRUE}, then unused levels will be pruned
from the resulting table}
}
\value{
A subset of \code{network} containing only sites
}
\description{
This function extracts the sites from a data.frame that contains both
sites and species (e.g. a data.frame read from \code{\link{readInfomapTree}})
}
\examples{



}
\author{
Boris Leroy \email{leroy.boris@gmail.com}
}
