% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupNodesPerCluster.R
\name{groupNodesPerCluster}
\alias{groupNodesPerCluster}
\title{Simplify the network into major clusters}
\usage{
groupNodesPerCluster(
  db,
  site.field = 1,
  species.field = 2,
  network,
  level = "lvl1",
  directed = FALSE,
  color.field = NULL,
  abundance.field = NULL
)
}
\arguments{
\item{db}{the database that was used to construct the network.}

\item{site.field}{name or number of column of sites}

\item{species.field}{name or number of column of species}

\item{network}{the network data.frame, usually issued from
\code{\link{readInfomapTree}}}

\item{level}{a character string specifying the level that will be used to
group nodes}

\item{directed}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, then the graph
will be directed in order to show the number of species/individuals in
 cluster Y that
come from cluster X and the reverse. If \code{FALSE}, then the graph will
simply show the number of species/individuals that are shared between both
clusters.}

\item{color.field}{the name of the color field associated to \code{level}}

\item{color}{the .tree file written by infomap.}
}
\value{
A data.frame
}
\description{
This function transforms the bipartite network data.frame into a unipartite
network data.frame that contains only major clusters. The link weights
correspond to the number of species shared between clusters or, in the case
where abundance is specified, the number of individuals between them.
}
\examples{
## NOT RUN
readInfomapTree("file.tree")

}
\author{
Boris Leroy \email{leroy.boris@gmail.com}
}
