% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participationCoefficient.R
\name{participationCoefficient}
\alias{participationCoefficient}
\title{Calculates the Participation Coefficient of nodes in a network}
\usage{
participationCoefficient(network, db, site.field, species.field, lvl = "lvl1")
}
\arguments{
\item{network}{the network data.frame, usually issued from
\code{\link{readInfomapTree}}}

\item{db}{the database that was used to construct the network.}

\item{site.field}{name or number of column of sites}

\item{species.field}{name or number of column of species}

\item{lvl}{column name of the level that you want to analyse}
}
\value{
The same data.frame as \code{network} with an additional column
containing participation coefficient values
}
\description{
This function calculates the participation coefficient, of nodes in a
network, i.e. the degree to which they are connected to nodes of a distinct
module. This metric is useful to identify, for example, transition zones,
which are zones that contain species from multiple regions.
}
\examples{



}
\author{
Boris Leroy \email{leroy.boris@gmail.com}
}
