% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGDF.R
\name{writeGDF}
\alias{writeGDF}
\title{Write a biogeographical network in GDF (Gephi) format}
\usage{
writeGDF(
  db,
  network = NULL,
  site.field = colnames(db)[1],
  species.field = colnames(db)[2],
  filename,
  color.field = NULL,
  abundance.field = NULL,
  hex2rgb = TRUE,
  directed = FALSE,
  additional.fields = NULL
)
}
\arguments{
\item{db}{a data.frame with at least two columns, site and species}

\item{network}{a network data.frame, typically obtained from
\code{\link{readInfomapTree}}}

\item{site.field}{name or number of site column}

\item{species.field}{name or number of species column}

\item{filename}{name of the GDF file to be written on disk. Extension
should be ".gdf".}

\item{color.field}{name of the column in \code{network}from which colors
should be fetched.
In Gephi, only one color column is used, so the desired one should be
specified here.}

\item{abundance.field}{(optional) name of abundance column}

\item{hex2rgb}{\code{TRUE} or {FALSE}. Specificies whether color codes should
be converted to RGB format. Required for Gephi 0.9.0-0.9.2}

\item{directed}{\code{TRUE} or {FALSE}. Specificies whether the network is
oriented or not.}

\item{additional.fields}{Not yet implemented.}
}
\description{
This function writes a biogeographical network in GDF file format
(readable by Gephi) with the possibility to add custom fields such as colors.
}
\examples{

}
\author{
Boris Leroy \email{leroy.boris@gmail.com}
}
