% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePajek.R
\name{writePajek}
\alias{writePajek}
\title{Write a biogeographical network in Pajek format}
\usage{
writePajek(
  db,
  site.field = 1,
  species.field = 2,
  filename,
  abundance.field = NULL
)
}
\arguments{
\item{db}{a data.frame with at least two columns, site and species. An
abundance field can also be useful.}

\item{site.field}{name or number column of sites}

\item{species.field}{name or number of column of species}

\item{filename}{name of the pajek file to be written on disk. Extension
should be ".net".}

\item{abundance.field}{(optional) name of abundance column}
}
\description{
This function writes a biogeographical network in pajek file format
(readable by infomap) on the basis of a data.frame with at least two
columns (sites and species).
}
\examples{
database <- data.frame(sites = c("site1", "site1", "site2", "site2"),
                       species = c("A", "B", "A", "C"),
                       abundance = c(5, 2, 7, 1))
 database

# Without abundance
writePajek(database, site.field = "sites", species.field = "species",
           filename = "pajektest.net")

# With abundance
writePajek(database, site.field = "sites", species.field = "species",
           filename = "pajektest.net", abundance.field = "abundance")

}
\author{
Boris Leroy \email{leroy.boris@gmail.com}
}
