% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeColors.R
\name{attributeColors}
\alias{attributeColors}
\title{Attribute colors to network clusters}
\usage{
attributeColors(
  network,
  nb.max.colors = 12,
  palette = "Paired",
  lvl = "lvl1",
  colname = "color",
  sh.grey = FALSE,
  other.color = grey(0.5),
  cluster.order = "sites",
  db,
  site.field = 1,
  species.field = 2
)
}
\arguments{
\item{network}{the network data.frame, usually issued from
\code{\link{readInfomapTree}} or getSitesTable or getSpeciesTable.}

\item{nb.max.colors}{a numeric <= 12 specifying the maximum number of
clusters to be colored.}

\item{palette}{a character string specifying the palette to use. See
\code{\link[RColorBrewer]{RColorBrewer}}}

\item{lvl}{a character string specifying the level to which you want to
attribute colors}

\item{colname}{a character string specifying how to name the color column}

\item{sh.grey}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, then all clusters
beyond \code{nb.max.colors} will be attributed distinct shades of grey. If
\code{FALSE}, they will all have the color specified at \code{other.color}}

\item{other.color}{a character string or function specifying the color to
attribute to clusters beyond \code{nb.max.colors}}

\item{cluster.order}{"undefined", "sites", "species", "sites+species".
This parameter defines in what order colors are attributed to clusters:
default order ("undefined"), or according to number of sites per cluster
 ("sites"), number of species per cluster ("species") or the total number
 of sites and species per cluster ("sites+species")}

\item{db}{a data.frame with at least two columns, site and species}

\item{site.field}{name or number of site column}

\item{species.field}{name or number of species column}
}
\value{
The same data.frame as \code{network} with an additional color
column
}
\description{
This function adds a color column to a network data.frame using RColorBrewer
palettes. A maximum number of clusters to colorize can be specified to only
highlight the most important clusters.
}
\examples{



}
\author{
Boris Leroy \email{leroy.boris@gmail.com}
}
