% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTables.R
\name{getSpeciesTable}
\alias{getSpeciesTable}
\title{Get the data.frame of species from a bipartite data.frame}
\usage{
getSpeciesTable(db, species.field = 2, network)
}
\arguments{
\item{db}{the database that was used to construct the network.}

\item{species.field}{a character string giving the name of species column in
\code{db}}

\item{network}{the network data.frame, usually read from
\code{\link{readInfomapTree}}}
}
\value{
A subset of \code{network} containing only species
}
\description{
This function extracts the species from a data.frame that contains both
sites and species (e.g. a data.frame read from \code{\link{readInfomapTree}})
}
\examples{



}
\author{
Boris Leroy \email{leroy.boris@gmail.com}
}
